KOGITO-1609: Tests
