+++ drools-wb-screens/drools-wb-scenario-simulation-editor/drools-wb-scenario-simulation-editor-client/src/test/java/org/drools/workbench/screens/scenariosimulation/client/editor/ScenarioSimulationEditorPresenterTest.java
@@ -727,74 +727,6 @@ public class ScenarioSimulationEditorPresenterTest extends AbstractScenarioSimul
         verify(scenarioSimulationEditorWrapperMock, never()).populateDocks(anyString());
     }
 
-    /* //TODO to move
-    @Test
-    public void populateRightDocks_Settings() {
-        when(presenterSpy.getCurrentRightDockPlaceRequest(anyString())).thenReturn(placeRequestMock);
-        doReturn(Optional.of(settingsPresenterMock)).when(presenterSpy).getSettingsPresenter(eq(placeRequestMock));
-        presenterSpy.populateRightDocks(SettingsPresenter.IDENTIFIER);
-        verify(presenterSpy, times(1)).setSettings(eq(settingsPresenterMock));
-        verify(settingsPresenterMock, times(1)).setCurrentPath(eq(pathMock));
-        verify(presenterSpy, never()).setTestTools(any());
-        verify(presenterSpy, never()).setCoverageReport(any());
-        verify(coverageReportPresenterMock, never()).setCurrentPath(any());
-    }
-
-    @Test
-    public void populateRightDocks_TestTools() {
-        when(presenterSpy.getCurrentRightDockPlaceRequest(anyString())).thenReturn(placeRequestMock);
-        doReturn(Optional.of(testToolsPresenterMock)).when(presenterSpy).getTestToolsPresenter(eq(placeRequestMock));
-        presenterSpy.populateRightDocks(TestToolsPresenter.IDENTIFIER);
-        verify(presenterSpy, never()).setSettings(any());
-        verify(settingsPresenterMock, never()).setCurrentPath(any());
-        verify(presenterSpy, times(1)).setTestTools(eq(testToolsPresenterMock));
-        verify(cheatSheetPresenterMock, never()).setCurrentPath(any());
-        verify(presenterSpy, never()).setCheatSheet(any());
-        verify(presenterSpy, never()).setCoverageReport(any());
-        verify(coverageReportPresenterMock, never()).setCurrentPath(any());
-    }
-
-    @Test
-    public void populateRightDocks_CheatSheetPresenter_NotShown() {
-        when(presenterSpy.getCurrentRightDockPlaceRequest(anyString())).thenReturn(placeRequestMock);
-        doReturn(Optional.of(testToolsPresenterMock)).when(presenterSpy).getTestToolsPresenter(eq(placeRequestMock));
-        presenterSpy.populateRightDocks(CheatSheetPresenter.IDENTIFIER);
-        verify(presenterSpy, never()).setSettings(any());
-        verify(settingsPresenterMock, never()).setCurrentPath(any());
-        verify(presenterSpy, never()).setTestTools(any());
-        verify(cheatSheetPresenterMock, times(1)).setCurrentPath(pathMock);
-        verify(presenterSpy, times(1)).setCheatSheet(eq(cheatSheetPresenterMock));
-        verify(presenterSpy, never()).setCoverageReport(any());
-        verify(coverageReportPresenterMock, never()).setCurrentPath(any());
-    }
-
-    @Test
-    public void populateRightDocks_CheatSheetPresenter_IsShown() {
-        when(cheatSheetPresenterMock.isCurrentlyShow(pathMock)).thenReturn(true);
-        when(presenterSpy.getCurrentRightDockPlaceRequest(anyString())).thenReturn(placeRequestMock);
-        doReturn(Optional.of(testToolsPresenterMock)).when(presenterSpy).getTestToolsPresenter(eq(placeRequestMock));
-        presenterSpy.populateRightDocks(CheatSheetPresenter.IDENTIFIER);
-        verify(presenterSpy, never()).setSettings(any());
-        verify(settingsPresenterMock, never()).setCurrentPath(any());
-        verify(presenterSpy, never()).setTestTools(any());
-        verify(cheatSheetPresenterMock, never()).setCurrentPath(any());
-        verify(presenterSpy, never()).setCheatSheet(any());
-        verify(presenterSpy, never()).setCoverageReport(any());
-        verify(coverageReportPresenterMock, never()).setCurrentPath(any());
-    }
-
-    @Test
-    public void populateRightDocks_CoverageReportPresenter() {
-        presenterSpy.populateRightDocks(CoverageReportPresenter.IDENTIFIER);
-        verify(presenterSpy, never()).setSettings(any());
-        verify(settingsPresenterMock, never()).setCurrentPath(any());
-        verify(presenterSpy, never()).setTestTools(any());
-        verify(cheatSheetPresenterMock, never()).setCurrentPath(any());
-        verify(presenterSpy, never()).setCheatSheet(any());
-        verify(presenterSpy, times(1)).setCoverageReport(eq(coverageReportPresenterMock));
-        verify(coverageReportPresenterMock, times(1)).setCurrentPath(eq(pathMock));
-    }*/
-
     @Test
     public void getModelSuccessCallbackMethod() {
         scenarioGridWidgetSpy.selectAndFocus();
